+++ pkg/common/utils/utils.go
@@ -7,41 +7,10 @@ package utils
 import (
 	"context"
 
-	"github.com/hibiken/asynq"
 	"github.com/uptrace/bun"
 
-	asynqclient "github.com/gardener/inventory/pkg/clients/asynq"
 	asynqutils "github.com/gardener/inventory/pkg/utils/asynq"
 )
-
-// TaskConstructor is a function which creates and returns a new [asynq.Task].
-type TaskConstructor func() *asynq.Task
-
-// Enqueue enqueues the tasks produced by the given task constructors.
-func Enqueue(ctx context.Context, items []TaskConstructor, opts ...asynq.Option) error {
-	logger := asynqutils.GetLogger(ctx)
-	for _, fn := range items {
-		task := fn()
-		info, err := asynqclient.Client.Enqueue(task, opts...)
-		if err != nil {
-			logger.Error(
-				"failed to enqueue task",
-				"type", task.Type(),
-				"reason", err,
-			)
-			return err
-		}
-
-		logger.Info(
-			"enqueued task",
-			"type", task.Type(),
-			"id", info.ID,
-			"queue", info.Queue,
-		)
-	}
-
-	return nil
-}
 
 // LinkFunction is a function, which establishes relationships between models.
 type LinkFunction func(ctx context.Context, db *bun.DB) error
