+++ pkg/openstack/tasks/tasks.go
@@ -31,14 +31,14 @@ func HandleCollectAllTask(ctx context.Context, t *asynq.Task) error {
 	queue := asynqutils.GetQueueName(ctx)
 
 	// Task constructors
-	taskFns := []utils.TaskConstructor{
+	taskFns := []asynqutils.TaskConstructor{
 		NewCollectServersTask,
 		NewCollectNetworksTask,
 		NewCollectLoadBalancersTask,
 		NewCollectSubnetsTask,
 	}
 
-	return utils.Enqueue(ctx, taskFns, asynq.Queue(queue))
+	return asynqutils.Enqueue(ctx, taskFns, asynq.Queue(queue))
 }
 
 // HandleLinkAllTask is a handler, which establishes links between the various
